fix(minimal deps)
