+++ docker/test-cpu-local.sh
@@ -4,6 +4,7 @@ set -ex
 echo "CONFIG"
 
 PYTHON_VERSION=${PYTHON_VERSION:-3.6}
+PIP_DEPS=${PIP_DEPS:--e .[dev]}
 WITH_NEO4J=${WITH_NEO4J:-0}
 WITH_LINT=${WITH_LINT:-1}
 WITH_TYPECHECK=${WITH_TYPECHECK:-1}
@@ -23,7 +24,10 @@ then
     ( cd ../test/db/neo4j && ./launch.sh )
 fi
 
-docker-compose build --build-arg PYTHON_VERSION=${PYTHON_VERSION} test-cpu
+docker-compose build \
+    --build-arg PYTHON_VERSION=${PYTHON_VERSION} \
+    --build-arg PIP_DEPS="${PIP_DEPS}" \
+    test-cpu
 
 echo "RUN"
